working time schedule domain

+++ src/main/java/com/durys/jakub/leaveentitlementsservice/workingtime/WorkingTimeSchedule.java
@@ -2,4 +2,4 @@ package com.durys.jakub.leaveentitlementsservice.workingtime;
 
 import java.math.BigDecimal;
 
-public record WorkingTimeSchedule(Integer days, BigDecimal hours) {}
+public record WorkingTimeSchedule(Long days, BigDecimal hours) {}
